www/requirements.txt
Django==2.2.28
django-cors-headers==2.4.1
djangorestframework==3.12.2
django-filter==2.4.0
djangorestframework-serializer-extensions==2.0.1